# spec/swagger.yaml
components:
  schemas:
...
    Commit:
      properties:
        commit:
          $ref: '#/components/schemas/GitCommit'
        html_url:
          type: string
        parents:
          items:
            $ref: '#/components/schemas/CommitParent'
          type: array
          uniqueItems: false
        sha:
          type: string
      required:
      - commit
      - html_url
      - parents
      - sha
      type: object
    CommitDiff:
      properties:
        diff:
          type: string
      required:
      - diff
      type: object
    CommitParent:
      properties:
        sha:
          type: string
      required:
      - sha
      type: object
...
    ErrorCode:
      enum:
      - bad_request
      - unauthenticated
      - not_found
      - method_not_allowed
      - internal
      - unavailable
      - review_not_found
      - idempotency_conflict
      - rate_limited
      - invalid_credentials
      - email_taken
      - invalid_email
      - weak_password
      - api_key_not_found
      - session_not_found
      - review_running
      - origin_forbidden
      - repository_not_found
      - provider_not_linked
      - provider_expired
      type: string
      x-enum-varnames:
      - CodeBadRequest
      - CodeUnauthenticated
      - CodeNotFound
      - CodeMethodNotAllowed
      - CodeInternal
      - CodeUnavailable
      - CodeReviewNotFound
      - CodeIdempotencyConflict
      - CodeRateLimited
      - CodeInvalidCredentials
      - CodeEmailTaken
      - CodeInvalidEmail
      - CodeWeakPassword
      - CodeAPIKeyNotFound
      - CodeSessionNotFound
      - CodeReviewRunning
      - CodeOriginForbidden
      - CodeRepositoryNotFound
      - CodeProviderNotLinked
      - CodeProviderExpired
    GitCommit:
      properties:
        author:
          $ref: '#/components/schemas/GitUser'
        committer:
          $ref: '#/components/schemas/GitUser'
        message:
          type: string
        tree:
          $ref: '#/components/schemas/TreeRef'
      required:
      - message
      - tree
      type: object
    GitTree:
      properties:
        sha:
          type: string
        tree:
          items:
            $ref: '#/components/schemas/GitTreeEntry'
          type: array
          uniqueItems: false
        truncated:
          type: boolean
      required:
      - sha
      - tree
      - truncated
      type: object
    GitTreeEntry:
      properties:
        mode:
          type: string
        path:
          type: string
        sha:
          type: string
        size:
          type: integer
        type:
          example: blob
          type: string
      required:
      - mode
      - path
      - sha
      - type
      type: object
    GitUser:
      properties:
        date:
          type: string
        email:
          type: string
        name:
          type: string
      type: object
...
    Installation:
      properties:
        account:
          $ref: '#/components/schemas/InstallationAccount'
        html_url:
          example: https://github.com/settings/installations/1
          type: string
        id:
          type: integer
        repository_selection:
          enum:
          - all
          - selected
          type: string
        target_type:
          example: User
          type: string
      required:
      - html_url
      - id
      - repository_selection
      - target_type
      type: object
    InstallationAccount:
      properties:
        login:
          example: octo
          type: string
      required:
      - login
      type: object
...
    Problem:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          example: review_not_found
        detail:
          example: review r_123 does not exist
          type: string
        request_id:
          type: string
        status:
          example: 404
          type: integer
        title:
          example: Not Found
          type: string
        type:
          example: about:blank
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
    Repository:
      properties:
        default_branch:
          example: main
          type: string
        full_name:
          example: octo/app
          type: string
        html_url:
          type: string
        name:
          example: app
          type: string
        owner:
          $ref: '#/components/schemas/RepositoryOwner'
        private:
          type: boolean
      required:
      - default_branch
      - full_name
      - html_url
      - name
      - owner
      - private
      type: object
    RepositoryList:
      properties:
        installations:
          items:
            $ref: '#/components/schemas/Installation'
          type: array
          uniqueItems: false
        items:
          items:
            $ref: '#/components/schemas/Repository'
          type: array
          uniqueItems: false
      required:
      - installations
      - items
      type: object
    RepositoryOwner:
      properties:
        login:
          example: octo
          type: string
      required:
      - login
      type: object
...
    TreeRef:
      properties:
        sha:
          type: string
      required:
      - sha
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Browser session set by sign-in
      in: cookie
      name: eyeful_session
      type: apiKey
...
info:
  description: Multi-agent code review with executable evidence.
  title: eyeful API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /repositories:
    get:
      description: Every installation of the GitHub App the user can access, with
        every repository of each. An installation's html_url is where the user changes
        its repositories.
      operationId: listRepositories
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service Unavailable
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: The GitHub repositories the signed-in user gave eyeful
      tags:
      - repositories
  /repositories/{owner}/{name}:
    get:
      description: The fields of GitHub's full-repository schema the console needs,
        read with the user's GitHub App token.
      operationId: getRepository
      parameters:
      - description: Owner
        example: octo
        in: path
        name: owner
        required: true
        schema:
          type: string
      - description: Name
        example: app
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service Unavailable
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: A GitHub repository the signed-in user gave eyeful
      tags:
      - repositories
  /repositories/{owner}/{name}/commits/{ref}:
    get:
      description: The fields of GitHub's commit schema the console needs, without
        its files.
      operationId: getRepositoryCommit
      parameters:
      - description: Owner
        example: octo
        in: path
        name: owner
        required: true
        schema:
          type: string
      - description: Name
        example: app
        in: path
        name: name
        required: true
        schema:
          type: string
      - description: Commit sha or branch
        example: main
        in: path
        name: ref
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commit'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service Unavailable
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: A commit, by sha or branch
      tags:
      - repositories
  /repositories/{owner}/{name}/commits/{ref}/diff:
    get:
      operationId: getRepositoryCommitDiff
      parameters:
      - description: Owner
        example: octo
        in: path
        name: owner
        required: true
        schema:
          type: string
      - description: Name
        example: app
        in: path
        name: name
        required: true
        schema:
          type: string
      - description: Full commit sha
        in: path
        name: ref
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitDiff'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service Unavailable
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: The unified diff a commit made against its first parent
      tags:
      - repositories
  /repositories/{owner}/{name}/git/trees/{sha}:
    get:
      description: GitHub's git-tree schema. truncated is true when GitHub returned
        only part of a large tree.
      operationId: getRepositoryTree
      parameters:
      - description: Owner
        example: octo
        in: path
        name: owner
        required: true
        schema:
          type: string
      - description: Name
        example: app
        in: path
        name: name
        required: true
        schema:
          type: string
      - description: Full tree sha
        in: path
        name: sha
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitTree'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service Unavailable
      security:
      - SessionCookie: []
      - BearerAuth: []
      summary: A git tree, recursively
      tags:
      - repositories
...
servers:
- url: /